fix: Address economic integrity leaks in sales tax and inheritance handlers
- Modified GoodsTransactionHandler and PublicManagerTransactionHandler to explicitly cast sales tax intents to integers, guaranteeing atomicity and strict integer math during settlement.
- Updated InheritanceHandler to distribute exactly tx.total_pennies rather than dynamically querying the agent's wallet balance, eliminating leaks from shared wallets or race conditions.
- Updated InheritanceManager to explicitly cast assets to float prior to rounding for precise tax calculations.
- Implemented Mock bool_ class in pytest conftest.py to resolve test runner collection failures associated with missing numpy dependencies.
- Added supplementary test scripts and JSON scenarios to verify zero-sum integrity continuously.
- Resolved multiple test failures induced by strict protocol compliance across DeathSystem, SagaOrchestrator, SettlementSecurity, and phase execution contexts.

+++ simulation/systems/handlers/public_manager_handler.py
@@ -38,13 +38,13 @@ class PublicManagerTransactionHandler(ITransactionHandler):
             # TD-233: Calculate Tax Atomically
             intents = context.taxation_system.calculate_tax_intents(tx, buyer, pm, context.government, context.market_data)
 
-            credits: List[Tuple[Any, float, str]] = []
-            credits.append((pm, trade_value, f"public_sale:{tx.item_id}"))
+            credits: List[Tuple[Any, int, str]] = []
+            credits.append((pm, int(trade_value), f"public_sale:{tx.item_id}"))
 
             # Note: total_cost calculation removed as it's not used here, settle_atomic handles total debit.
 
             for intent in intents:
-                credits.append((context.government, intent.amount, intent.reason))
+                credits.append((context.government, int(intent.amount), intent.reason))
 
             success = context.settlement_system.settle_atomic(buyer, credits, context.time)
